pyproject.toml
[project]
name = "working-capital-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.5",
    "fastapi>=0.135.1",
    "google-generativeai>=0.8.6",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "sentence-transformers>=5.3.0",
    "uvicorn>=0.42.0",
]
